+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of lizard-fewsnorm
 ===================================================
 
 
-0.3 (unreleased)
+0.3 (2011-11-28)
 ----------------
 
 - Fixed sync_fewsnorm after removing method FewsNormSource.o.